Refactored utils/index. Grouped services by actions. Each service has it's own place now
